move PathInfo to pkg package
