+++ io/pkg/gflows_package.go
@@ -1,0 +1,8 @@
+package pkg
+
+type GFlowsPackage interface {
+	WorkflowsDir() string
+	LibsDir() string
+	GetPathInfo(localPath string) (*PathInfo, error)
+	// Validate?
+}
